Pull out watch list to its own component
